refactor: split store.tsx into store/ directory
