+++ src/renderer/lib/store/index.ts
@@ -1,0 +1,4 @@
+export type { LensView, AppState, AppAction } from './state';
+export { initialState } from './state';
+export { getPlainContent, handleChatEvent, appReducer } from './reducer';
+export { AppStateProvider, useAppState, useAppDispatch } from './context';
